+++ CHANGES.md
@@ -3,6 +3,8 @@
 ### Release 4.3.2
 * Fixed wrong Font Awesome in css
 * Type collapsedtext now use Moodle css classes to show the carets for collapsed and expanded.
+* Introduce Drag-and-Drop for element with items like "accordion", "carousel" and "grid"
+* Introduce adding and removing of elements by ajax
 
 ### Release 4.3.1
 * mod_unilabel: prevent loading activities like inline folder (#36)